- Added the Buy module. - Corrected score types based on what the Continuum client displays.   + Kills and Deaths (AKA Wins and Losses) are UInt16. Updated the S2C_PlayerData packet and the S2C_ScoreUpdate packet.   + Kill points and flag points are Int32. Unfortunately, I didn't know negative values were allowed and previously used a UInt64 to store the actual stat (larger type so that the "forever" interval won't overflow). Had to change it to Int64, so added a hacky compatibilty fix to workaround reading the wrong type from existing stats databases. - Corrected the speed personal best stat type to be signed, so that it matches what Continuum displays for the S2C_SpeedStats packet. - Updated the Stats module to use Generic Math which was added in .NET 7. DateTime does not implement INumber<T> so separated it out as a TimestampStatInfo. Also, removed the methods for incrementing the DateTime stats them since it's ambiguous on whether the initial value should be DateTime.MinValue or the DateTime.UtcNow.

+++ src/Core/Modules/Game.cs
@@ -1965,7 +1965,7 @@ namespace SS.Core.Modules
                 {
                     try
                     {
-                        allPlayerStats.IncrementStat(killer, StatCodes.KillPoints, null, (ulong)points);
+                        allPlayerStats.IncrementStat(killer, StatCodes.KillPoints, null, points);
                     }
                     finally
                     {
